Added input check for html option
Since I was totally owned by the fact that this is should to a BOOL, not the html that you want to set. This is completely contrary to intuition (because we are in jQuery land where you set html using "$(el).html(myHtmlStringHere)", so I suspect a lot more people than me will make this mistake. Right now, the tooltip just doesn't appear, no error or nothing - incredibly bad usability.